+++ CHANGELOG.md
@@ -14,6 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 
 - [Patch drupal/core to fix Issue #3414883: datetime_timestamp widget does not use default field value #771](https://github.com/farmOS/farmOS/pull/771)
+- [Fix duplicated revision tab on entities #773](https://github.com/farmOS/farmOS/pull/773)
 
 ## [3.0.0] 2024-01-05